Refactor Lora card view model
